+++ src/Calabonga.Commandex.Engine/Base/Commands/ParameterCommandexCommand.cs
@@ -1,5 +1,4 @@
-﻿using Calabonga.Commandex.Engine.Base;
-using Calabonga.Commandex.Engine.Exceptions;
+﻿using Calabonga.Commandex.Engine.Exceptions;
 using Calabonga.Commandex.Engine.Extensions;
 using Calabonga.Commandex.Engine.Settings;
 using Calabonga.OperationResults;
@@ -7,7 +6,7 @@ using System.IO;
 using System.Text;
 using System.Text.Json;
 
-namespace Calabonga.Commandex.Engine.Commands;
+namespace Calabonga.Commandex.Engine.Base.Commands;
 
 /// <summary>
 /// // Calabonga: Summary required (CommandexCommand 2024-07-29 09:38)
@@ -53,6 +52,11 @@ public abstract class ParameterCommandexCommand<TParams> : ICommandexCommand
     public abstract Task<OperationEmpty<ExecuteCommandexCommandException>> ExecuteCommandAsync();
 
     public object? GetResult() => Parameter;
+
+    /// <summary>
+    ///  Tags about what the command is intended for
+    /// </summary>
+    public abstract string[] Tags { get; }
 
     protected TParams? Parameter
     {
